creating new package.json in package dir during search build
