+++ package.json
@@ -24,19 +24,14 @@
     "start:storybook": "start-storybook -p 6006",
     "build:storybook": "build-storybook",
     "clean": "rimraf searchdist",
-    "build:search": "npm run clean && rollup -c --bundleConfigAsCjs",
-    "prepublish": "rm -rf ./searchdist && npm run build:search",
-    "build:package": "tsc -p ./package_scripts/tsconfig.json --outDir ./package && tsc -p ./package_scripts/tsconfig.json && node ./scripts/createPackage.js",
-    "prepare": "npm run builduipackage"
+    "buildsearch": "npm run clean && rollup -c --bundleConfigAsCjs",
+    "prepublish": "rm -rf ./searchdist && npm run buildsearch",
+    "buildpackage": "tsc -p ./package_scripts/tsconfig.json --outDir ./package && tsc -p ./package_scripts/tsconfig.json && node ./scripts/createPackage.ts",
+    "prepare": "npm run buildpackage"
   },
   "files": [
-    "src/**/*",
-    "searchdist/**/*"
+    "src/**/*"
   ],
-  "typings": "searchdist/index.d.ts",
-  "typescript": {
-    "definition": "searchdist/index.d.ts"
-  },
   "dependencies": {
     "@ant-design/charts": "^1.3.4",
     "@ant-design/icons": "^4.4.0",
@@ -73,6 +68,7 @@
     "deep-object-diff": "^1.1.0",
     "express": "^4.16.4",
     "express-prom-bundle": "^5.0.2",
+    "fs-extra": "^10.1.0",
     "git-revision-webpack-plugin": "^3.0.6",
     "handlebars": "^4.7.7",
     "history": "^4.7.2",
@@ -144,6 +140,7 @@
     "@types/cookie-parser": "^1.4.1",
     "@types/enzyme": "^3.9.0",
     "@types/express": "^4.16.1",
+    "@types/fs-extra": "^9.0.13",
     "@types/jest": "^24.0.11",
     "@types/jquery": "^3.5.14",
     "@types/jquery-mousewheel": "^3.1.9",
