+++ package_scripts/index.d.ts
@@ -1,0 +1,3 @@
+declare const directoryPath: string;
+declare const defaultFilePath: string;
+export { directoryPath, defaultFilePath };
